requirements.txt
fastapi[all]~=0.70.1
psycopg2
sqlalchemy~=1.4.29
pydantic~=1.9.0
passlib[bcrypt]
python-jose[cryptography]